Update issue templates
Https://oscarg933.github.io.patch.diff.md
Https://codecov.io.patch.diff.md .

Choose an endpoint
 
Introduction
The api.video webservice helps you put video on the web without the hassle. Just upload your content and we'll do (everything) else...

API conventions
Base URL
The api.video endpoint is https://ws.api.video where ws stands for web service.

Authentication
The api.video webservice uses oAuth authentication protocol. See  section for all the details.

Naming
Use hyphenated-lower-case for query Path Segments.
Use lowerCamelCase for query parameters, as they will appear in JSON keys.
Sorting
sortBy and sortOrder optional parameters control result sorting.

sortBy value must be a resource property, and sortOrder must be asc or desc.

Properties that are eligible to sorting are listed in documentation for each resource.

Actions
Some responses provide a set of actions.

Actions define what can be done on a resource or resource collection. They allow easy fine-grained frontend integration instead of dealing with project permissions / channel roles / owner privileges.

For example:

If project_update action is present, you can display the Edit button.
If channel_view_statistics action is present, you can display the Statistics panel.
Available actions on a given response depend on resource type, your global permissions, your channel roles and your resource ownership.

Available actions are defined in each endpoint documentation.

API evolution
API uses content negotiation to address a specific API version:

Accept: application/vnd.api.video+json;version=1;
The Vary header ensures different versions of the same resource are cached independently.

Only the major version is negotiable.

By default, the last version is served. It means that the API works in a browser where it always serves the latest stable version:

Internally, the API uses  versions that are shown in every HTTP response:

Api-Version: 1.0
Pagination
Several services return item collections and need pagination.

You can paginate a collection with parameters page and pageSize in a query:

GET /channels?page=3&pageSize=10
page is the requested page number (computed according to pageSize). It can be any positive integer. If the requested page exceeds the size of the collection, a HTTP 200 status is returned with an empty collection.
pageSize is the expected number of items. It's a positive integer between 1 and 100. The default is 25. A HTTP 418 response is returned for invalid values.
Paginated responses contain 3 sections:

{
    "data": [ {..}, {..}, {..} ],
    "actions": ["..", ".."],
    "pagination": {
        "itemsTotal": 17,
        "pagesTotal": 6,
        "pageSize": 3,
        "currentPage": 4,
        "currentPageItems": 3,
        "links": [
          {
            "rel": "first",
            "uri": "/xyz?pageSize=3"
          },
          {
            "rel": "previous",
            "uri": "/xyz?page=3&pageSize=3"
          },
          {
            "rel": "current",
            "uri": "/xyz?page=4&pageSize=3"
          },
          {
            "rel": "next",
            "uri": "/xyz?page=5&pageSize=3"
          },
          {
            "rel": "last",
            "uri": "/xyz?page=6&pageSize=3"
          }
        ]
    }
}
data section is an array containing the items of the page
actions section, as detailed above.
pagination section:
itemsTotal: total number of items in the entire collection
pagesTotal: total number of pages containing pageSize items in the entire collection
pageSize: number of items per page
currentPage: current page
currentPageItems: number of items in the current page (it may be pageSize, or less if the current page is the last one)
links to current, first, previous, next and last pages when available
When collections fit in a single page, the pagination section is returned anyway:

{
    "data": [ {..}, {..}, {..} ],
    "actions": ["..", ".."],
    "pagination": {
        "itemsTotal": 3,
        "pagesTotal": 1,
        "pageSize": 25,
        "currentPage": 1,
        "currentPageItems": 3,
        "links": [
          {
            "rel": "current",
            "uri": "/xyz?pageSize=3"
          }
        ]
    }
}
Problem details
This API follows . When several problems occur, an extra problems property is added, containing all of them.

For example:

{
  "type": "https://docs.api.video/problems/parameter.out_of_list",
  "title": "This parameter is out of the list of allowed values.",
  "parameter": "status",
  "allowed": [
    "scheduled",
    "imminent",
    "live",
    "ended"
  ],
  "problems": [
    {
      "type": "https://docs.api.video/problems/parameter.out_of_list",
      "title": "This parameter is out of the list of allowed values.",
      "parameter": "status",
      "allowed": [
        "scheduled",
        "imminent",
        "live",
        "ended"
      ]
    },
    {
      "type": "https://docs.api.video/problems/parameter.out_of_list",
      "title": "This parameter is out of the list of allowed values.",
      "parameter": "sortBy",
      "allowed": [
        "start",
        "end",
        "status"
      ]
    }
  ]
}
Problem details list
https://docs.api.video/problems/forbidden
You are not allowed to perform the requested operation.

https://docs.api.video/problems/parameter.invalid
This parameter is invalid.

https://docs.api.video/problems/parameter.out_of_list
This parameter is out of the list of allowed values. These values are available in documentation and in response contents, under allowed key.

https://docs.api.video/problems/parameter.out_of_range
This parameter is out of the allowed range of values. The range property indicates the allowed range with optional min and max values.

https://docs.api.video/problems/parameter.required
This parameter is required.

https://docs.api.video/problems/property.datetime.invalid
The datetime format is invalid. Datetimes must respect  standard, for example 1977-04-22T06:00:00Z.

https://docs.api.video/problems/property.invalid
This property is invalid.

https://docs.api.video/problems/property.out_of_list
This property is out of the list of allowed values. These values are available in documentation and in response contents, under allowed key.

https://docs.api.video/problems/property.required
This property is required.

https://docs.api.video/problems/resource.not_found
The requested resource is not found. This might happen when the resource has been deleted, or the current user doesn't have enough credentials.

1